Progress Bar and Refactoring
Added a progress bar to the view. This will become visible with a
specific loading message when a database activity is called. I've kept
it in thee main view rather than opening a new dialog as JavaFX doesn't
seem to have good support for a progress dialog box, and I don't know
how long these database actions will take. It wouldn't be reasonable to
hold up the entire UI for 30 mins with a slow database and a dodgy
connection, or a user loading 1000's of files.