Fixed JDBC queries and Encryptor
Idiot errors in the JDBC - using executeQuery instead of executeUpdate.

Found a serious flaw in my encryption assumptions. An encryptor with a
completely different algorithm would be created each time the object was
instantiated or the system was restarted, rendering database decryption
impossible.

Changed it to be a thread safe singleton and manually set a basic key.
This now works with a restarted system.